% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_input_file_ex.R
\docType{data}
\name{excel_input_file_ex}
\alias{excel_input_file_ex}
\title{Multi-state model analysis results file}
\format{
An object of class \code{data.frame} with 303 rows and 211 columns.
}
\usage{
excel_input_file_ex
}
\description{
Example csv file for the structure of the results input file for the multi-state setting illness-death model of 3 states (EBMT example).
\describe{
\item{Ntransitions}{This variable will have the same number for each row, that is, the number of transitions (e.g 3)}
\item{atlist}{ The value of this variable specify the name of the covariate pattern for which the prediciton is made in the specific row.}
\item{timevar}{ It is the time variable. The value of this variable specify the time point for which the prediciton is made for a certain covariate pattern.}
\item{P(_diff/_ratio)_X_to_X(_uci/_lci)}{Transition Probabilities for each state}
\item{Haz(_diff/_ratio)_X_to_X(_uci/_lci)}{"Haz" stands for the hazard or transition rate foreach transition.}
\item{Los(_diff/_ratio)_X_to_X(_uci/_lci)}{"Los" stands for Length of stay".}
\item{Visit(_diff/_ratio)_X_to_X(_uci/_lci)}{"Visit" stands for the "Probability of ever being" in each state.}
\item{User(_diff/_ratio)_X_to_X(_uci/_lci)}{"User" stands for the "User specified function".}
\item{Next_X_to_X(_uci/_lci)}{"Next" stands for the "Probability that a particular state r is the next state after the current state s".}
\item{Number_X_to_X(_uci/_lci)}{"Number" stands for the expected number of visits to a state.}
\item{First_X(_uci/_lci)}{"First" stands for the expected "First passage time" into a particular state.}
\item{Soj_X(_uci/_lci)}{"Soj" stands for the "mean sojourn time" into a particular state.}
\item{uci stands for "Upper confidence interval". The upper confidence interval of the specified measure is stored in variables with this ending.}{}
\item{lci stands for "Lower confidence interval". The lower confidence interval of the specified measure is stored in variables with this ending.}{}
\item{diff stands for "Difference". The difference in the specified measure between the present covariate pattern and the covariate pattern of reference is stored in this variable.}{}
\item{ratio stands for "Ratio". The ratio of the specified measure between the present covariate pattern and the covariate pattern of reference is stored in this variable.}{}
\item{X_to_X: The first asterisk specifies the starting state at time s and the second asterisk specifies the landing state at time t. If a measure has only one asterisk then it is a measure that refers to a specific state that does not entail a sense of transition between states.}{}
}
}
\keyword{datasets}
