% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSMplus-package.R
\docType{package}
\name{MSMplus-package}
\alias{MSMplus}
\alias{MSMplus-package}
\title{MSMplus: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
