% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMSMplus.R
\name{runMSMplus}
\alias{runMSMplus}
\title{Local deployment of MSMplus application}
\usage{
runMSMplus()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Launches MSMplus locally. We advice to click at "Open in a Browser" option
of the window that will open. It works best with the Google Chrome Browser
}
\details{
Details on how to use the MSMplus app can be found in the app platform.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE
 runMSMplus()
 }
}
}
\seealso{
\code{\link[shiny]{runApp}}
}
