% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSMplus-package.R
\docType{package}
\name{MSMplus-package}
\alias{MSMplus}
\alias{MSMplus-package}
\title{MSMplus: Prepare MSMplus input files and locally deploys the MSMplus app}
\description{
Prepare input files and local deployment of MSMplus. The MSMplus package contains four functions for the automatic derivativation of the json input files for the MSMplus webtool (msboxes_R(), msmjson(),flexsurvjson() and mstatejson()). It also contains function runMSMplus() for the local deployment of the app, runMSMplus_help() which opens a help page that describes the web tool and runMSMplus_sup() which opens a tutorial on how to build the json/csv input files for the MSMplus app. To have access to the vignette, the user needs to type install_github("nskourlis/MSMplus", build_vignettes = TRUE, dependencies=TRUE) and after calling the MSMplus library, then type browseVignettes("MSMplus").
}
\author{
\strong{Maintainer}: Nikolaos Skourlis \email{nikolaos.skourlis@ki.se} (\href{https://orcid.org/0000-0002-5003-9611}{ORCID})

}
\keyword{internal}
