% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSMplus-package.R
\docType{package}
\name{MSMplus-package}
\alias{MSMplus}
\alias{MSMplus-package}
\title{MSMplus: Prepare MSMplus input files and locally deploys the MSMplus app}
\description{
Prepare input files and local deployment of MSMplus.
}
\author{
\strong{Maintainer}: Nikolaos Skourlis \email{nikolaos.skourlis@ki.se} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
