% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMSMplus.R, R/runMSMplus_help.R,
%   R/runMSMplus_sup.R
\name{runMSMplus}
\alias{runMSMplus}
\alias{runMSMplus_help}
\alias{runMSMplus_sup}
\title{Local deployment of MSMplus application}
\usage{
runMSMplus()

runMSMplus_help()

runMSMplus_sup()
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
Launches MSMplus locally. We advice to click at "Open in a Browser" option
of the window that will open. It works best with the Google Chrome Browser

Launches the  MSMplus help page locally. We advice to click at "Open in a Browser" option
of the window that will open. It works best with the Google Chrome Browser

This page is a tutorial on how to build the json/csv input files for the MSMplus app. It also contains examples in Stata and R of
conducting the multi-state analysis and deriving the neccesary input files.
of the window that will open. It works best with the Google Chrome Browser
}
\details{
Details on how to use the MSMplus app can be found in the app platform.

Details on how to use the MSMplus application.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE
 runMSMplus()
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE
 runMSMplus_help()
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE
 runMSMplus_sup()
 }
}
}
\seealso{
\code{\link[shiny]{runApp}}

\code{\link[shiny]{runApp}}

\code{\link[shiny]{runApp}}
}
